backend/app/utils/jwt_utils.py
"""
JWT 工具函数
"""
from datetime import datetime, timedelta
from typing import Optional
from jose import jwt, JWTError
from app.config import settings


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    """
    创建访问令牌
    
    Args:
        data: 要编码的数据
        expires_delta: 过期时间增量
        
    Returns:
        JWT token 字符串
    """
    to_encode = data.copy()
    
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    
    to_encode.update({"exp": expire})
    
    encoded_jwt = jwt.encode(
        to_encode,
        settings.SECRET_KEY,
        algorithm=settings.ALGORITHM
    )
    
    return encoded_jwt


def verify_access_token(token: str) -> Optional[dict]:
    """
    验证访问令牌
    
    Args:
        token: JWT token 字符串
        
    Returns:
        解码后的数据，验证失败返回 None
    """
    try:
        payload = jwt.decode(
            token,
            settings.SECRET_KEY,
            algorithms=[settings.ALGORITHM]
        )
        return payload
    except JWTError:
        return None


def get_token_expire_time(token: str) -> Optional[datetime]:
    """
    获取令牌过期时间
    
    Args:
        token: JWT token 字符串
        
    Returns:
        过期时间，验证失败返回 None
    """
    payload = verify_access_token(token)
    if payload and "exp" in payload:
        return datetime.utcfromtimestamp(payload["exp"])
    return None
